[EASI-2410] Ask a Question Modal Functionality in Read Only pages (#576)
* Extracted Question and Reply into its own component

* blowup read only discussions

* Add back the reply parts

* Add back the reply parts

* move teh conditional inside the onclick

* remove some commented code

* add a bunch of optional props

* embracing the QuestionAndReply component

* pass modal state setter down

* sort abc

* remove weird logic
